Fix compilation and migrate code to VS 2019
The previous code drop had a few issues and the code didn't compile. But
that was enough for people who just wanted to take a look at the code.

With this commit:
- Migrate project file from VS Express 2008 to VS Community 2019
- Add two missing files
- Disable _TV
- Fix compilation (some compiler flags had to be changed)

The result is that I can build the code in debug mode.
Compilation in release mode needs a bit more work (some link issues).